requirements.txt
python-box
meshtastic
paho-mqtt
asyncio-paho
pyserial
aiofiles
